Moved graphql_table and graphqlquery
