+++ psrdb/graphql_table.py
