+++ psrdb/scripts/psrdb.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-from psrdb.tables.graphql_table import GraphQLTable
+from psrdb.graphql_table import GraphQLTable
 from psrdb.graphql_client import GraphQLClient
 
 from psrdb.tables.main_project import MainProject
