+++ psrdb/tables/ephemeris.py
@@ -1,8 +1,8 @@
 import hashlib
 import json
 
-from psrdb.tables.graphql_table import GraphQLTable
-from psrdb.tables.graphql_query import graphql_query_factory
+from psrdb.graphql_table import GraphQLTable
+from psrdb.graphql_query import graphql_query_factory
 
 
 class Ephemeris(GraphQLTable):
