+++ psrdb/tables/observation.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
-from psrdb.tables.graphql_table import GraphQLTable
-from psrdb.tables.graphql_query import graphql_query_factory
+from psrdb.graphql_table import GraphQLTable
+from psrdb.graphql_query import graphql_query_factory
 
 
 class Observation(GraphQLTable):
