+++ psrdb/tables/pipelinefiles.py
@@ -1,7 +1,7 @@
 import os
 
-from psrdb.tables.graphql_table import GraphQLTable
-from psrdb.tables.graphql_query import graphql_query_factory
+from psrdb.graphql_table import GraphQLTable
+from psrdb.graphql_query import graphql_query_factory
 from base64 import b64encode
